Split packet trimming casee into symmetric and asymmetric
