+++ tests/generic_config_updater/conftest.py
@@ -171,3 +171,18 @@ def ignore_expected_loganalyzer_exceptions(duthosts, selected_dut_hostname, loga
             ".*ERR ctrmgrd.py: Join failed.*"
         ]
         loganalyzer[duthost.hostname].ignore_regex.extend(ignoreRegex)
+
+
+@pytest.fixture(scope="session")
+def skip_if_packet_trimming_not_supported(duthost):
+    """
+    Check if the current device supports packet trimming feature.
+    """
+    platform = duthost.facts["platform"]
+    logger.info(f"Checking packet trimming support for platform: {platform}")
+
+    # Check if the SWITCH_TRIMMING_CAPABLE capability is true
+    trimming_capable = duthost.command('redis-cli -n 6 HGET "SWITCH_CAPABILITY|switch" "SWITCH_TRIMMING_CAPABLE"')[
+        'stdout'].strip()
+    if trimming_capable.lower() != 'true':
+        pytest.skip("Packet trimming is not supported")
